+++ src/mykrobe/parser.py
@@ -181,8 +181,12 @@ parser_amr.add_argument(
     "--conf_percent_cutoff",
     metavar="conf_percent_cutoff",
     type=float,
-    help="Number between 0 and 100. Determines --min_variant_conf, by simulating variants and choosing the cutoff that would keep x%% of the variants. Default is 90 if --ont, otherwise --min_variant_conf is used as the cutoff",
-    default=-1,
+    help=(
+        "Number between 0 and 100. Determines --min_variant_conf, by simulating "
+        "variants and choosing the cutoff that would keep x%% of the variants "
+        "(default: %(default)d)"
+    ),
+    default=100,
 )
 parser_amr.add_argument(
     "-O",